Added model class for subrddit and mapped it to communityType

+++ src/main/java/com/reddit/RedditClone/repository/SubredditRepository.java
@@ -3,6 +3,6 @@ package com.reddit.RedditClone.repository;
 import com.reddit.RedditClone.model.Subreddit;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubredditRepository extends JpaRepository<Subreddit, Integer> {
+public interface SubredditRepository extends JpaRepository<Subreddit, Long> {
 
 }
